Fiks sirkelavhengighet
