+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/brev/DokumentServiceTest.kt
@@ -9,7 +9,7 @@ import no.nav.familie.ba.sak.config.AbstractSpringIntegrationTestDev
 import no.nav.familie.ba.sak.config.ClientMocks
 import no.nav.familie.ba.sak.config.DatabaseCleanupService
 import no.nav.familie.ba.sak.config.TEST_PDF
-import no.nav.familie.ba.sak.integrasjoner.journalføring.JournalføringService
+import no.nav.familie.ba.sak.integrasjoner.journalføring.UtgåendeJournalføringService
 import no.nav.familie.ba.sak.kjerne.arbeidsfordeling.ArbeidsfordelingService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
@@ -68,7 +68,7 @@ class DokumentServiceTest(
     private val brevService: BrevService,
 
     @Autowired
-    private val journalføringService: JournalføringService,
+    private val utgåendeJournalføringService: UtgåendeJournalføringService,
 
     @Autowired
     private val arbeidsfordelingService: ArbeidsfordelingService,
@@ -267,7 +267,7 @@ class DokumentServiceTest(
         dokumentService.sendManueltBrev(manueltBrevRequest, behandling, behandling.fagsak.id)
 
         io.mockk.verify(exactly = 1) {
-            journalføringService.journalførManueltBrev(
+            utgåendeJournalføringService.journalførManueltBrev(
                 fnr = manueltBrevRequest.mottakerIdent,
                 fagsakId = behandling.fagsak.id.toString(),
                 journalførendeEnhet = any(),
